Add DialogBox
Show dialog if the user wants to remove a file which is currently being uploaded

+++ src/app/tabs/filelist/FileItem.tsx
@@ -1,10 +1,11 @@
 import {observer}        from 'mobx-react';
 import {Component, h}    from 'preact';
-import {files}           from '../../../state';
+import {files, uploads}  from '../../../state';
 import {ListedFile}      from '../../../state/models/ListedFile';
 import {copyToClipboard} from '../../../utils/copy-to-clipboard';
 import {bind, cn}        from '../../../utils/preact-utils';
 import {isMobile}        from '../../browserenv';
+import {DialogBox}       from '../../overlays/DialogBox';
 import {Toast}           from '../../overlays/Toast';
 import {FileStatus}      from './FileStatus';
 import styles            from './FileItem.module.scss';
@@ -23,7 +24,7 @@ export class FileItem extends Component<Props, State> {
     @bind
     copyLink() {
         const {id, file} = this.props.item;
-        const toast = Toast.getInstance();
+        const toast = Toast.instance;
         const link = `${env.API_ENDPOINT}/d/${id}`;
 
         /**
@@ -54,7 +55,34 @@ export class FileItem extends Component<Props, State> {
         const {id} = this.props.item;
 
         if (id) {
-            files.removeFile(id);
+            const relatedUploads = uploads.listedUploads.filter(v => v.listedFile.id === id).length;
+
+            // Tell the user that uploads are about to get cancelled
+            if (relatedUploads > 0) {
+                DialogBox.instance.open({
+                    icon: 'exclamation-mark',
+                    title: 'Uh Oh! Are you sure about that?',
+                    description: relatedUploads > 1 ?
+                        `There are currently ${relatedUploads} uploads for this file. Continue?` :
+                        'This file is currently being uploaded. Continue?',
+                    buttons: [
+                        {
+                            type: 'success',
+                            text: 'Keep File'
+                        },
+                        {
+                            type: 'error',
+                            text: 'Remove'
+                        }
+                    ]
+                }).then(value => {
+                    if (value === 1) {
+                        files.removeFile(id);
+                    }
+                });
+            } else {
+                files.removeFile(id);
+            }
         }
 
         e.stopPropagation();
